fix($srn): remaining stories bugs

+++ srn/scripts/wasm-build.js
@@ -3,10 +3,7 @@ const fs = require('fs-extra');
 const isWin = process.platform === 'win32';
 const yargs = require('yargs');
 
-async function buildForWeb({
-  noTransform,
-  transformOnly,
-}) {
+async function buildForWeb({ noTransform, transformOnly, noCleanTmp }) {
   if (!transformOnly) {
     console.log('Building rust code for extracting TS definitions...');
     await spawnWatched(
@@ -15,7 +12,7 @@ async function buildForWeb({
         spawnOptions: {
           cwd: 'world',
         },
-      },
+      }
     );
     const mainBuildCommand =
       'wasm-bindgen target/wasm32-unknown-unknown/debug/world.wasm --typescript --out-dir pkg-bindgen';
@@ -25,7 +22,7 @@ async function buildForWeb({
         spawnOptions: {
           cwd: 'world',
         },
-      },
+      }
     );
     await fs.move('world/pkg-bindgen/world.d.ts', 'world/world.d.ts.tmp', {
       overwrite: true,
@@ -45,10 +42,10 @@ async function buildForWeb({
     });
     console.log('Running codemods...');
     await spawnWatched(
-      'yarn jscodeshift -t codeshift/make-smart-enums.ts --extensions=ts world/pkg/world.d.ts',
+      'yarn jscodeshift -t codeshift/make-smart-enums.ts --extensions=ts world/pkg/world.d.ts'
     );
     await spawnWatched(
-      'yarn jscodeshift -t codeshift/make-builder-classes.ts --extensions=ts world/pkg/world.d.ts',
+      'yarn jscodeshift -t codeshift/make-builder-classes.ts --extensions=ts world/pkg/world.d.ts'
     );
     const file = (await fs.readFile('world/pkg/world.d.ts')).toString();
     const builderClassFinder = /\/\/ start builder class (\w+)(?:.|\n|\r)*export class \1(?:.|\n|\r)*end builder class \1/gm;
@@ -66,10 +63,22 @@ async function buildForWeb({
     await fs.writeFile('world/pkg/world.d.ts', cleanedFile + extraImportsFile);
     const builders = `type Uuid = string; \n${extractedBuilders.join('\n\n')}`;
     const enums = extractedEnums.join('\n\n');
+    // this can, of course, be generated, but I'm lazy =)
+    const extraExtraImports = `
+    import { Vec2f64, ObjectSpecifier, ManualMoveUpdate, LongActionStart,
+    InventoryAction, NotificationActionR, SandboxCommand, TradeAction, NotificationText, ShootTarget,
+    SBAddPlanet, SBTeleport, SBSetupState
+    } from "./world"
+    `;
     console.log('writing the extra file');
-    await fs.writeFile('world/pkg/world.extra.ts', `${enums}\n\n${builders}`);
-    console.log('deleting original tmp file...');
-    await fs.unlink('world/world.d.ts.tmp');
+    await fs.writeFile(
+      'world/pkg/world.extra.ts',
+      `${extraExtraImports}\n\n${enums}\n\n${builders}`
+    );
+    if (!noCleanTmp) {
+      console.log('deleting original tmp file...');
+      await fs.unlink('world/world.d.ts.tmp');
+    }
     console.log('Done, ts definitions + wasm binary are ready!');
   }
 }
@@ -82,7 +91,7 @@ async function buildForTests() {
       spawnOptions: {
         cwd: 'world',
       },
-    },
+    }
   );
   const mainBuildCommand =
     'wasm-bindgen target/wasm32-unknown-unknown/debug/world.wasm --no-modules --out-dir pkg-nomodule';
@@ -93,7 +102,7 @@ async function buildForTests() {
       spawnOptions: {
         cwd: 'world',
       },
-    },
+    }
   );
   console.log('Patching code...');
   let file = (await fs.readFile('world/pkg-nomodule/world.js')).toString();
@@ -107,12 +116,12 @@ async function buildForTests() {
   // new version of require-patching for jest in wasm-bindgen 0.2.79 (may differ for other versions)
   file = file.replace(
     'var ret = getObject(arg0).require(getStringFromWasm0(arg1, arg2));',
-    'var ret = require(getStringFromWasm0(arg1, arg2)); // patch for wasm-bindgen+jest',
+    'var ret = require(getStringFromWasm0(arg1, arg2)); // patch for wasm-bindgen+jest'
   );
 
   if (file.indexOf('patch for wasm-bindgen+jest') === -1) {
     throw new Error(
-      'wasm-bindgen require patch has failed, the require calls from rust won\'t work!',
+      "wasm-bindgen require patch has failed, the require calls from rust won't work!"
     );
   }
 
@@ -131,7 +140,7 @@ export const getBindgen = () => {
   console.log('Done, file is ready!');
 }
 
-(async function() {
+(async function () {
   yargs
     .command(
       '$0',
@@ -145,21 +154,23 @@ export const getBindgen = () => {
           .option('transformOnly', {
             type: 'boolean',
             default: false,
+          })
+          .option('noCleanTmp', {
+            type: 'boolean',
+            default: false,
           }),
-      async ({
-        noTransform,
-        transformOnly,
-      }) => {
+      async ({ noTransform, transformOnly, noCleanTmp }) => {
         try {
           await buildForWeb({
             noTransform,
             transformOnly,
+            noCleanTmp,
           });
         } catch (e) {
           console.error(e);
           process.exit(1);
         }
-      },
+      }
     )
     .command(
       'forTests',
@@ -172,7 +183,7 @@ export const getBindgen = () => {
           console.error(e);
           process.exit(1);
         }
-      },
+      }
     )
     .parse();
 })();
